fix(sheet): make Import one control with a menu, the way Export already is
The toolbar carried two upload icons — "Import {type}" and "Import survey responses" — which
read as two unrelated actions, and the first silently depended on the data source type set in
the Settings tab: importing a JSON file meant opening Settings, switching Type, coming back.
Export sitting right beside it already had the right shape, one button whose menu names the
format.

Import is now that same control: the menu lists the formats, and picking one reads the file
with that format's default settings when it is not the sheet's own — the fallback the export
dialog already makes. Survey responses is the last item under a divider, since it is a source
rather than a file format, and its dialog no longer carries a button of its own.
